Specify target to fix old babel issue. Improve errors.
If a user had a tsconfig.json file with a target like es5, esbuild picks up that file and may have issues transpiling the code to that older format. An alternative workaround would be to pre-transform with babel, but since we want to reduce dependencies and we should be recommending Node >= 14, we can transpiled to a ES2020 target and it should work nicely in all Node env's 14 or greater.

See: https://github.com/evanw/esbuild/issues/1010